requirements.txt


google-generativeai[adk]